+++ src/test/java/com/klarna/hiverunner/sql/cli/CommandShellEmulationTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.klarna.hiverunner.sql;
+package com.klarna.hiverunner.sql.cli;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.Matchers.isEmptyString;
