+++ src/test/java/com/klarna/hiverunner/sql/split/StatementsSplitterTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.klarna.hiverunner.sql;
+package com.klarna.hiverunner.sql.split;
 
 import java.util.Arrays;
 import java.util.List;
